requirements.txt
python-telegram-bot
Flask
psycopg2-binary
google-generativeai
gunicorn